refactor: remove as-any casts, extract entity services for storage and telemetry
