+++ apps/server-core/src/adapters/http/controllers/entities/analysis-node-log/handlers/create.ts
@@ -14,7 +14,7 @@ import { LogChannel, LogFlag } from '@privateaim/telemetry-kit';
 import type { Request, Response } from 'routup';
 import { sendAccepted } from 'routup';
 import { useDataSource } from 'typeorm-extension';
-import { HTTPHandlerOperation, useRequestIdentityRealm } from '@privateaim/server-http-kit';
+import { useRequestIdentityRealm } from '@privateaim/server-http-kit';
 import { RoutupContainerAdapter } from '@validup/adapter-routup';
 import { ValidupError, buildErrorMessageForAttribute, defineIssueItem } from 'validup';
 import {
@@ -26,7 +26,7 @@ import { AnalysisNodeLogValidator } from '../utils/index.ts';
 export async function createAnalysisNodeLogRouteHandler(req: Request, res: Response, telemetryClient?: TelemetryClient) : Promise<any> {
     const validator = new AnalysisNodeLogValidator();
     const validatorAdapter = new RoutupContainerAdapter(validator);
-    const data = await validatorAdapter.run(req, { group: HTTPHandlerOperation.CREATE });
+    const data = await validatorAdapter.run(req, { group: 'create' });
 
     const dataSource = await useDataSource();
     const nodeRepository = dataSource.getRepository(NodeEntity);
