+++ apps/server-storage/src/app/components/bucket/handlers/create/module.ts
@@ -14,11 +14,10 @@ import {
 import { DomainType } from '@privateaim/storage-kit';
 import { LogFlag } from '@privateaim/telemetry-kit';
 import { useDataSource } from 'typeorm-extension';
-import { HTTPHandlerOperation } from '@privateaim/server-http-kit';
 import type { Client } from 'minio';
 import { BucketEntity } from '../../../../../adapters/database/index.ts';
 import { toBucketName } from '../../../../domains/bucket/utils.ts';
-import { BucketValidator } from '../../../../../adapters/http/controllers/bucket/utils/validation.ts';
+import { BucketValidator, ValidatorGroup } from '../../../../../core/entities/index.ts';
 
 export class BucketCreateHandler implements ComponentHandler<
     BucketComponentEventMap,
@@ -71,7 +70,7 @@ export class BucketCreateHandler implements ComponentHandler<
             value,
         );
 
-        const data = await this.validator.run(value, { group: HTTPHandlerOperation.CREATE });
+        const data = await this.validator.run(value, { group: ValidatorGroup.CREATE });
 
         const dataSource = await useDataSource();
         const repository = dataSource.getRepository(BucketEntity);
