+++ apps/server-storage/src/core/utils/index.ts
@@ -5,4 +5,5 @@
  * view the LICENSE file that was distributed with this source code.
  */
 
+export * from './bucket-name.ts';
 export * from './stream-to-buffer.ts';
